# openapi.yaml
openapi: 3.0.3
info:
  title: rustdesk-panel-api
  description: >-
    RustDesk 管理平台后端契约（单一契约源，oapi-codegen design-first）。
    M1 交付 healthz + 认证域 + 用户自身端点共 25 个 operation；
    M2 交付核心业务域 43 个 operation（设备端协议 heartbeat/sysinfo、
    设备/设备组、策略、RBAC、用户组）；M3 交付 96 个 operation
    （用户域 CRUD/邀请/批量/安全 14、通讯录 32、审计 9、仪表盘 2、
    服务器管理 10、nexus 9、设置 9、OIDC 提供者 8、更新检查 1、
    静态资源 2），总计 164 个 operation。
    错误响应统一 NestJS 包络 {statusCode, message, error}，
    message 三种形态并存（业务对象 / 纯文本 / 校验字符串数组）；
    POST 成功一律 200。限流 per-IP per-route（超限 429 ThrottlerException）；
    设备端协议为设备维度限流（tracker=id→uuid→IP）。
    契约基准：参考项目主分支实测（附录 A/B），仅取契约、不搬实现。
  version: 1.1.0-m2
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/login-options:
    get:
      summary: 登录方式选项（OIDC 提供商列表；参考源码实测为 GET）
      description: 限流 20 次/分钟/IP。返回 ["oidc/{name}"] 或 [{name, icon}]。
      operationId: loginOptions
      tags: [oidc]
      responses:
        "200":
          description: 登录选项数组
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: object
                      required: [name]
                      properties:
                        name:
                          type: string
                        icon:
                          type: string
        default:
          $ref: "#/components/responses/Error"
...
  /api/oidc/callback:
    get:
      summary: OIDC 提供商回调（HTML 页，客户端按 rememberMe 存储 token 后跳转）
      operationId: oidcCallback
      tags: [oidc]
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 回调 HTML 页面
          content:
            text/html:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
...
  /api/heartbeat:
    post:
      summary: 设备心跳上报（无鉴权；设备维度限流 10 次/分钟，tracker=id→uuid→IP）
      description: >-
        未知设备自动注册（首跳 INSERT ON CONFLICT UPDATE）。响应条件键：
        有 pending 断连时含 disconnect；解析出策略且 strategy.updatedAt(毫秒)
        严格大于报文 modified_at 时含 strategy.config_options 与新 modified_at；
        均未命中时为 {}。未知字段 400（forbidNonWhitelisted 语义）。
      operationId: heartbeat
      tags: [device]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: 心跳确认（条件键响应，可为空对象）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: 设备维度限流
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/sysinfo:
    post:
      summary: 设备系统信息上报（无鉴权；设备维度限流 5 次/分钟，tracker=id→uuid→IP）
      description: >-
        按 uuid 查 peers，不存在直接返回 ID_NOT_FOUND（不自动注册）；存在则
        upsert sysinfos（核心字段"提供即覆盖"、preset 列"非空真值才覆盖"），
        并处理 preset-device-group-name 关联与 preset-note 兜底写入；
        preset-address-book-* 字段接受但不处理（M3 通讯录联动）。
        响应恒为 text/plain 且 HTTP 200。
      operationId: sysinfo
      tags: [device]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SysinfoRequest"
      responses:
        "200":
          description: "SYSINFO_UPDATED（设备已注册并更新）或 ID_NOT_FOUND（设备未注册）"
          content:
            text/plain:
              schema:
                type: string
                enum: [SYSINFO_UPDATED, ID_NOT_FOUND]
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: 设备维度限流
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"

  # ==================== M2 设备域（peers 同表两视图） ====================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    HeartbeatRequest:
      type: object
      additionalProperties: false
      required: [id, uuid, ver, modified_at]
      properties:
        id:
          type: string
          description: RustDesk 数字设备 ID（如 "123456789"）
        uuid:
          type: string
          description: 设备 UUID（peers 主键）
        ver:
          type: integer
          description: 客户端整版本号（格式化规则见 info.version）
        modified_at:
          type: integer
          format: int64
          description: 客户端策略版本戳（毫秒 epoch）
        conns:
          type: array
          items:
            type: integer
          description: 当前活跃连接 ID；键缺失 ≠ 空（指针判空区分）
    HeartbeatResponse:
      type: object
      additionalProperties: false
      properties:
        disconnect:
          type: array
          items:
            type: integer
          description: 待断连 connId 列表（客户端确认语义）
        strategy:
          type: object
          required: [config_options]
          properties:
            config_options:
              type: object
              additionalProperties:
                type: string
        modified_at:
          type: integer
          format: int64
          description: 下发用策略 updatedAt（毫秒）
    SysinfoRequest:
      type: object
      additionalProperties: false
      required: [uuid]
      properties:
        uuid:
          type: string
        id:
          type: string
        hostname:
          type: string
        username:
          type: string
        os:
          type: string
        cpu:
          type: string
        memory:
          type: string
        version:
          type: string
        preset-address-book-name:
          type: string
          description: M2 接受不处理（M3 通讯录联动）
        preset-address-book-tag:
          type: string
        preset-address-book-alias:
          type: string
        preset-address-book-password:
          type: string
        preset-address-book-note:
          type: string
        preset-username:
          type: string
          description: 非空真值才覆盖 sysinfos.preset_username
        preset-strategy-name:
          type: string
        preset-device-group-name:
          type: string
          description: 按名关联 device_groups 并回写 peers.deviceGroupGuid
        preset-note:
          type: string
          description: 仅当 peers.note 为空时写入
...
    Error:
      type: object
      description: NestJS 默认错误包络；message 为对象/字符串/字符串数组三种形态之一。
      required: [statusCode, message, error]
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
            - type: object
        error:
          type: string
  responses:
    BadRequest:
      description: 参数或校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Error:
      description: 错误包络
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...